% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bankomat}
\alias{Bankomat}
\title{ATM withdrawals.}
\format{A data frame with 239060 rows and 15 variables:
\describe{
  \item{ymd}{year month and day of withdrawal}
  \item{weekday}{weekday of withdrawal}
  ...
}}
\source{
\url{http://www.trutschnig.net/Datensatz.txt}
}
\usage{
Bankomat
}
\description{
A dataset containing the sum of withdrawal and the corresponding weekday
including a marker for holidays.
}
\keyword{datasets}
