% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstFunction.R
\name{firstFunction}
\alias{firstFunction}
\title{firstFunction}
\usage{
firstFunction(x, y)
}
\arguments{
\item{x}{a scalar or vector}

\item{y}{a scalar or vector}
}
\value{
returns the sum of the two parameters
}
\description{
firstFunction sums two inputs
}
