% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso12}
\alias{iso12}
\title{iso12 provides regions.}
\format{A data frame with 239060 rows and 15 variables:
\describe{
  \item{iso_adm2}{district iso number}
  \item{district}{name of the district}
  ...
}}
\source{
\url{http://www.trutschnig.net/iso12.RData}
}
\usage{
iso12
}
\description{
A dataset containing the download and upload speed and other attributes of
about 239,000 mobile users.
}
\keyword{datasets}
