% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RTR.R
\name{PlotRTR}
\alias{PlotRTR}
\title{Gives a scatterplot}
\usage{
PlotRTR(region)
}
\arguments{
\item{region}{}
}
\value{
a scatterplot of RTR by region
}
\description{
Gives a scatterplot
}
\examples{
PlotRTR("Salzburg")
}
