% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTR}
\alias{RTR}
\title{RTR download and upload speed with regions.}
\format{A data frame with 248349 rows and 20 variables:
\describe{
  \item{id}{id of user}
  \item{mtyme}{time at upload or download}
  ...
}}
\usage{
RTR
}
\description{
A dataset containing the download and upload speed and other attributes of
about 239,000 mobile users.
}
\keyword{datasets}
