% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTR2015}
\alias{RTR2015}
\title{RTR download and upload speed.}
\format{A data frame with 239060 rows and 15 variables:
\describe{
  \item{id}{id of user}
  \item{mtyme}{time at upload or download}
  ...
}}
\source{
\url{http://www.trutschnig.net/RTR2015.RData}
}
\usage{
RTR2015
}
\description{
A dataset containing the download and upload speed and other attributes of
about 239,000 mobile users.
}
\keyword{datasets}
