% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstFunction.R
\name{secondFunction}
\alias{secondFunction}
\title{secondFunction}
\usage{
secondFunction(x, y)
}
\arguments{
\item{x}{a sclar}

\item{y}{a scalar}
}
\value{
the product ot two inputs
}
\description{
builds the product of two inputs
}
